[#60] feat: imgDownLoad 유틸
